菜单表完善

+++ common/src/main/java/me/zrxjava/common/base/TreeNode.java
@@ -17,10 +17,10 @@ import java.util.List;
 public class TreeNode {
 
 	@ApiModelProperty(value = "当前节点id")
-	protected Integer id;
+	protected Long id;
 
 	@ApiModelProperty(value = "父节点id")
-	protected Integer pid;
+	protected Long pid;
 
 	@ApiModelProperty(value = "子节点列表")
 	protected List<TreeNode> children = new ArrayList<>();
